Feat: 로그인매니저분리
